Don't use CLONE_SETTLS - it doesn't work as expected. Instead use "SIGCHLD | CLONE_VFORK".
This is sad news. We needed CLONE_SETTLS, so that the child process have a copy
of its own "errno", etc. variables. This way the child could run *in parallel*
with the parent process until it executes exec().

Unfortunately, CLONE_SETTLS immediately terminates the unit tests with SIGSEGV.
Something with it doesn't work as expected, or I don't understand how it work.

This forces us to use "CLONE_VFORK" which suspends the parent process until the
child executes exec(). And we need to suspend the parent, because we don't
have our copy of "errno", etc. and some of the calls inside the child could
modify it.

More or less, popen_noshell() uses an equivalent of vfork() now.

SIGCHLD is also required. Otherwise, when the child process terminates,
it is automatically reaped and we can't get its status in the parent process.
Since the parent is suspended during the execution (because of CLONE_VFORK),
the usage of SIGCHLD is now safe.